Update LLVM from 14.0.6 to 21.1.2 (#8737)
This PR upgrades slang-llvm to use LLVM 21.1.2. It also fixes the
`USE_SYSTEM_LLVM` option on some Linux distributions (e.g. SUSE and
derivatives) which don't ship separate components for LLVM or Clang, and
instead just have `libLLVM.so` and `libclang-cpp.so` containing all
components.

## Motivation

I'm working on a [direct LLVM IR
emitter](https://github.com/juliusikkala/slang/tree/llvm-target) for
Slang. That work is in a fairly good state by now, but I still need to
move the emitter into slang-llvm and link to LLVM in the Correct Way
instead of whatever garbage I have in that branch right now. The problem
with that is that slang-llvm currently links to a comparatively ancient
LLVM version. There have been major changes to LLVM since then, and I
wrote the emitter for LLVM 20-21. Differences between those two versions
are minor and I can downgrade to 20 with little effort if that's better.