Add neural.slang module (#8602)
Close #8570 and #8571.

This PR adds neural.slang into slang repo.

`neural.slang` will be part of our standard library to provide users
some utility to implement inline MLP.

### Source Location

The source of this module is located at `source/slang-neural-module/`,
the structure of the directory is as follow
```
source/standard-modules/
├── CMakeLists.txt
├── neural
│   ├── bindless-storage.slang
│   ├── buffer-storage.slang
│   ├── CMakeLists.txt
│   ├── inline-vector.slang
│   ├── istorages.slang
│   ├── ivector.slang
│   └── neural.slang
├── README.md
└── slang-standard-module-config.h.in
```

where `CMakeList.txt` is integrated into the main build process, such
that we can package the prebuilt slang module into our final release
package. Other than that, we will also package the source as well.

### Build Process
When launching the build, the cmake will first generate
`slang-standard-module-config.h` from
`slang-standard-module-config.h.in` according to the config defined in
`CMakeList.txt`, where we define the install directory and module name
in this header file so that our compiler will know which folder to look
for it.

Then it will compile all the sources into `neural.slang-module`, and
copy all the source code and prebuilt module into
`slang-prebuilt-module` directory that is at the same directory with
`slang.dll/libslang.so`.

### Module searching mechanism

We implement a new attribute [ExperimentalModule] to decorate a module.
Once a module is decorated by this attribute, we will only load this
module when `-experimental-feature` compiler flag is set.

When `-experimental-feature` is set and user module contains `import
neural;, the compiler should provide a way to find the location of the
module automatically.

We achieve this by first looking for the `slang.dll/libslang.so` being
used in the compilation. Since we will always package the module the
same directory as the c++ library, we will just search the module in
that directory. But if there is a user defined module named the same, we
will not load the builtin module.

This mechanism should work well when people have different version of
slang installed on their system, it will not have the compatible
problem, because the pre-built module must be compiled by the slang
compiler that is packaged at the same location.

### Unit Test
This is the first PR for the integration, so the full implementation and
test plan is not done yet. But we can already have some unit test to
test the basic data structure and algorithm. We will put all the unit
tests into `tests/neural` directory.