Save cache only if the job is running from master branch (#8864)
We should only save cache for the master branches, because the cache
from feature branches can't be restored and used by other branches, and
there is only limited amount of space.
If the cache from the master branch is evicted, we will hit cache miss
for all LLVM build and results in long building time.
This PRs adds this check.
But we should run CI on master branch periortically to make sure new
cache are generated. Right now it seems only the manually dispatched job
will run on master. Most of jobs post-merge are running in the
merge_queue.

Co-authored-by: Yong He <yonghe@outlook.com>