Rename library from slang to slang-compiler and version library names (#8746)
Fixes #4016, #7694

This change renames the base name of the "slang" library to
"slang-compiler" on all platforms.

For Windows, we provide:
- slang.dll, a proxy library that forwards all functions to
slang-compiler.dll (transitional)
- slang.lib (transitional)

On Linux and macOS, we provide a symlink from the old library name to
the new one. (transitional)

Also adds versioning to filenames of dynamically linked libraries for
Linux and macOS:
- libslang-compiler.{EXT}.0.{VERSION}
- libgfx.{EXT}.0.{VERSION}
- libslang-rt.{EXT}.0.{VERSION}

Runtime-loaded libraries for Linux and macOS are versioned as follows:
- libslang-glsl-module-{VERSION}.{EXT}
- libslang-glslang-{VERSION}.{EXT}

Items marked "transitional" will be removed in the future (ETA: end of
2026)

Includes documentation about the change.

---------

Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>