Be more picky about version tags
Narrow the scope of what we consider to be a valid version tag for our
versioning scheme, which means YEAR.NUM[.NUM[.NUM]]

The first part of this version tag is used for our library filename
versioning, so it needs to be valid for that schema.  Previously we
accepted anything that started with a 'v', which was overly broad.