Skip lowering SPIRV bitcasts that can use OpBitcast (#8361)
Fixes #7013

This changes the bitcast IR lower pass to check if the target is direct
emission SPIR-V, then if so, checks if the operand and result type are
natively supported by `OpBitcast` without lowering, and skips lowering
those cases.

This includes pointer<->pointer bitcasts of the same storage class,
pointer<->integer bitcasts, 32-bit integer vector<->pointer bitcasts,
and vector<->scalar and vector<->vector bitcasts where the operands are
the same size, and where the greater element count is an integer
multiple of the lesser element count.

---------

Co-authored-by: slangbot <ellieh+slangbot@nvidia.com>
Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>