[DO NOT MERGE] Add __supportImplicitDerivatives with the respective IR op
Add IR op for determining the support for implicit derivatives based
on whether the corresponding capability is available.

This commit **should not** be merged at this time. The reason being
that we should rather investigate on implementing a proper
`if constexpr (checkCapability(...))` or similar construction. In this
light, the IR op would only be a temporary clean-up. However, it is
somewhat unlikely that `__targetHasImplicitDerivatives()` would get
out of sync with the capability definitions, since implicit-LOD
sampling is a well-established feature.

Issue #8683