Fix crash when emitting DebugInlinedAt (#8881)
When emitting SPIRV, the 4th operand of kIROp_DebugInlinedAt is nullptr
with the repro shader. From the debugging, it turned out that Debug
related decorators were not properly copied into the synthesized
backward autodiff functions.

This PR includes the following IRs to the decorator list to copy:
 - kIROp_DebugLocationDecoration
 - kIROp_DebugFuncDecoration

This PR also fixes a bug in `fixUpDebugFuncType()`. It creates a new
`IRDebugFunction` with the new type and it was incorrectly attaching the
decoration to the type not the function.