add "slang.workspaceFlavor" to language server config, allowing selecting standard/vfx (#8915)
Slang supports VFX syntax by checking file extension. We use a forked
Source 2 where our format is the same but our extension is ".shader" and
our includes are ".hlsl" which also require `EnableEffectAnnotations`.

This adds the configurability to the language server which enables us to
offer full intellisense to users developing shaders.

Happy to make this work differently if it's not the way.