Fix slang's slangpy CI testing across versions
Relates to #8828

Whenever you're testing, say, a slang 2025.22 release using slangpy
built against slang 2025.21.2, you need to account for the library
name difference when attempting to overwrite the slang libraries in
the locally installed python site-package.

This change plugs a hole in testing that could occur in the time
between when a new slang release is made, and when an updated slangpy
release is pushed to PyPI that uses that new slang release.