Add minimal repro for issue 8920 and update SPIRV tools and headers (#8981)
The SPIRV tools optimizer had an issue with the aggressive dead code
elimination pass, which is exposed by the newly added minimal repro. In
practice, this issue either triggered the following assert:

slangc:
/builds3/slang/external/spirv-tools/source/opt/def_use_manager.cpp:56:
void
spvtools::opt::analysis::DefUseManager::AnalyzeInstUse(spvtools::opt::Instruction*):
Assertion 'def && "Definition is not registered."' failed.

or it crashed slangc when the asserts were disabled.

This issue is already fixed by
https://github.com/KhronosGroup/SPIRV-Tools/commit/917617c93734a94446332fe62aed1f90cddcd61a
in the upstream, so sync SPIRV-Tools and SPIRV-Headers as follows to
make this issue go away:

- SPIRV-Tools: 917617c93734a94446332fe62aed1f90cddcd61a
- SPIRV-Headers: f2e4bd213104fe323a01e935df56557328d37ac8

This commit does not upgrade SPIRV-Tools to the latest and greatest. A
later commit, likely
https://github.com/KhronosGroup/SPIRV-Tools/commit/6570fd4076d3f6ba7e1349af042d9f97f7378884
, triggers SPIR-V validation errors in test
language-feature/pointer/ptr-to-groupshared.slang. This is tracked by
issue #8989 .

Issue #8920