Fix documentation of __targetHasImplicitDerivatives()
Also, add capability `implicit_derivatives_sampling`. This should be a
requirement for functions such as _Texture.Sample() and
_Texture.CalculateLevelOfDetail(). But currently this cannot be
achieved without complex __target_switch/__stage_switch constructs in
GLSL texture() functions due to requirement inheritance.

Further, enforcing this requirements with static_assert() does not
allow capabilities to be inferred, which has too high a likelihood of
breaking existing code. So, we're not doing that in this patch,
either. Instead, we accept that we can generate invalid code without
diagnostics for now.

Issue #8683