Embed VERSION in source archives, give priority
Fixes #9019

Up until now, we've used git describe as the primary source of truth
for the version number. To be clear, it still is, but we need to
consider what downstream consumers and packagers of Slang may be
doing. e.g., the following use cases:

- Rebuilding from source archives

- Rebuilding from source archives ingested into a separate version
  control system

To that end, we define a VERSION file, and enable git export-subst on
it. For source releases created using git archive, this will replace
the contents of VERSION with the most recent version number, as
expressed relative to the most recent version tag in the Slang git
repository (our source of truth).

The cmake rule is updated to prioritize the VERSION file (if and only
if export-subst has been applied to the file) higher than the git
describe contents, since the source package may have been ingested
into another VCS (possibly also git). It's possible this won't be
accurate--additional patches may have been applied on top--but we're
doing what we can here to ensure the versioning on the filenames is
correct. Not every eventuality can be forseen.