Rename VERSION to SLANG_VERSION
Rename `VERSION` to `SLANG_VERSION` to avoid a macOS-specific build
failure.

On macOS, the filesystem is often configured as case-insensitive.
Well, C++ STL includes a file called `version`. Unfortunately, the
root of the repository is getting included for whatever reason, which
means that #include <version> picked up our `VERSION` file instead of
the `version` file it was expecting..