CI builds should fetch tags
Since Slang builds use git tags extracted via git describe as the
source of truth for version information, we should ensure the tags are
available whenever we build (both for CI and for Releases).

For release builds, this issue was obscured because we used the
triggering ref to determine the version. This worked, for the most
part, except for some internal versioning information not readily
visible, which may have been incorrect.

For CI builds, this appears to have been incorrect for some time, but
not necessarily an issue, since we don't normally need to have the
version numbers 'correct' in CI. In practice, though, we should
probably be using the same path for CI as for Releases so we don't
have diverging behavior.