Fix null pointer dereference in emitCalleeDebugInlinedAt (#8978)
It appears that there are cases where the IRDebugLocationDecoration and
IRDebugFuncDecoration are not present on the caller side when the
inlining happens.

The repro case was when the function is exported from another module and
it was a generic function.

I have seen that there were cases where we cannot print an error message
with the exact source location when the error handling happens much
later in the backend. I am guessing that this is probably the same case
where some of the debug information is missing when the specialization
from another module is handled much later in the backend.

This commit skips the debug-inline-info when there isn't enough
information to emit.