Improve typename deduction in slang-emit-spirv
In SPIRVEmitContext::getName(), add some logic to detect when inst is an
IRType*, and call getTypeNameHint() to see if it matches any of the
types listed in that function.

This helps in the case where inst refers to a
RaytracingAccelerationStructure. With this patch, the structure name
shows up correctly in the SPIR-V output; before this patch the structure
name shows up as "unamed".

Also, change RayTracingAccelerationStructure ->
RaytracingAccelerationStructure, to match the typename as it appears in
slang source code.

Add a test to verify that the typename appears as
RaytracingAccelerationStructure in the SPIR-V output.

Fixes #9025