Sanitize triggering ref
In the release workflows, we use the triggering ref as the source of
version information, when generating the package filenames. But the
triggering ref can be a branch name, and a common pattern is for these
branch names to have a hierarchy, e.g., "dev/USER/branchname". Hence
the sanitization of version information.