Fix generic function overloads in reflection API (#9052)
Close #8633.

When calling `findFunctionByName<someGeneric<concrete_type>>` to find a
generic function, it will report nullptr if there are overloads for the
generic function.

The reason is that we didn't handle OverloadedExpr2 node in the
reflection API.

The fix is to conservatively convert OverloadedExpr2 to OverloadedExpr
and reuse to existing logic to keep handling the OverloadedExpr.

The alternative solution will be add logic of handling OverloadedExpr2
to every location that only handles OverloadedExpr. But it might be more
complicated.