Fix shadow cube array sampling for GLSL target
In GLSL, shadow cube array requires a slightly different parameter
form in texture() and textureLod() functions. This is because in other
types, the location vector includes the usual sample location appended
with array index and the depth reference comparison value. However, in
shadow cube array, this vector would now grow to 5 elements whereas 4
is the maximum. Hence, the comparison value is a separate parameter.

Add also a regression test for this feature.

Issue #9074