Fix documentation of __targetHasImplicitDerivatives() (#9014)
Also, add capability `implicit_derivatives_sampling`. This should be a
requirement for functions such as _Texture.Sample() and
_Texture.CalculateLevelOfDetail(). But currently this requirement cannot
be added without complex __target_switch/__stage_switch constructs in
GLSL texture() functions because the calling function has to have at
least
the capabilities of the called function.
    
Further, enforcing the implicit derivatives requirement with
static_assert()
does not allow capabilities to be added on demand, which has too high
a likelihood of breaking existing code. Therefore, we'll accept that we
may
generate invalid code without proper diagnostics for now.

Issue #8683