Add post-fixup phase in rewriteValueUsesToAddrUses + regression test
Function rewriteValueUsesToAddrUses() translates
IRFieldExtract/IRGetElement/IRStore instructions that are affected by
the change of the function parameter type from value to
pointer. However, the implementation failed to consider the case where
the index of IRGetElement() is IRFieldExtract() in which case the
translated IRGetElementPtr() would be passed a pointer as the index.

Add a second phase in this function that fixes this issue by adding a
pointer load.

Add also a minimal repro for this case as a regression test.

Issue #9073