Add layout rule for ending CUDA entry point parameter structs without padding (#9090)
Fixes #9007

CUDA structs are required to be padded to their largest alignment. but
if you attempt to use the size of the entry point parameter struct to
determine the size of a kernel launch parameter buffer using
`TypeLayoutReflection::getSize()`, the extra padding will cause
`cuLaunchKernel` to fail, as the buffer is required to have no such
padding.

The entry point parameter struct is not emitted as a struct in CUDA, so
it does not need to follow the same alignment rule as a normal struct,
and one could expect that `getSize()` should return a size appropriate
for actual use in CUDA, so it should not have the padding.

This change adds a new layout rules implementation,
`CUDAEntryPointParameterLayoutRulesImpl`, that has a definition of
`EndStructLayout(UniformLayoutInfo* ioStructInfo)` that does not add the
trailing padding that
`CUDALayoutRulesImpl::EndStructLayout(UniformLayoutInfo* ioStructInfo)`
adds to the end of structs.

With this, you can derive the unpadded size for a parameter buffer from
an `EntryPointReflection *entryPointLayout` with
`entryPointLayout->getTypeLayout()->getElementTypeLayout()->getSize()`.

The change also adds a unit test to verify the lack of padding, and a
slang test that currently passes due to slang-rhi using a workaround for
the issue. I have verified that the slang test will continue to pass if
the method above is used for the parameter buffer size instead of the
workaround.

---------

Co-authored-by: slangbot <ellieh+slangbot@nvidia.com>
Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>