Correct inheritence documentation and add reminder to update docs (#9117)
Fixes #8930

This corrects the inheritance documentation to specify that it is
specifically _interface_ types that structs can inherit from, instead of
types in general, and adds a note there about the deprecated status of
struct inheritance.

This also adds a note to the CONTRIBUTING doc to remind to verify that
docs align with new/changed diagnostics (e.g. struct inheritance used to
be permitted, but now produces an error, so the doc should no longer
imply that as a valid case).