Add SLANG_VERSION for source archives (#9020)
Fixes #9019

Slang already depended on git tags for version information, but a recent
change to add library versioning created an additional dependency on git
to provide the library version number when generating the library
filenames. This information is simply not available in a source package.
(The source packages we generate don't configure correctly with the
preset profiles due to the submodules not being included in the
packaging, but that doesn't mean we should make the problem _worse_.)

To address this, a `cmake/slang_git_version` file has been added to
store the version number, which is filled out automatically on `git
archive` using git's `export-subst`. For source releases created using
`git archive`, this will replace the contents of `slang_git_version`
with the most recent version number.

`GitVersion.cmake` is updated to prioritize the contents of the
`slang_git_version` file (if and only if `export-subst` has been applied
to the file) higher than the `git describe` results. This order was
chosen in case the source archive was subsequently ingested into a
different git repository.

---------

Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>