Legalize matrix operations where only operand needs lowering (#9134)
Fixes #8823

In the matrix legalization pass, we currently do not legalize the
instruction if the result matrix type does not need lowering. This
causes an issue for cast operations, where the result and operand types
are typically not the same. If the result type is a matrix type that
does not need lowering (e.g. a matrix of `IRFloatType`) but the operand
is a matrix type that _does_ need lowering (e.g. a matrix of
`IRUintType`), then the instruction does not get legalized and the
operand matrix type is never lowered as a result.

This change adds a check for whether the operand matrix type needs
lowering, and legalizes matrix operations if _either_ the result or
operand need lowering.

This change also splits `legalizeUnaryOperation` into
`legalizeUnaryCastOperation` and `legalizeUnaryLogicalOperation`. Both
do the `SLANG_ASSERT` checks done by `legalizeUnaryOperation`, before
calling a modified `legalizeUnaryOperation` to do the rest of the
operation legalization, but `legalizeUnaryCastOperation` lacks the
assertion that the result type should need legalization as that is not
always true for casts as explained above.
The modified `legalizeUnaryOperation` emits an `IRMakeMatrix` if the
result matrix type does not need lowering, instead of always emitting an
`IRMakeArray`.

---------

Co-authored-by: slangbot <ellieh+slangbot@nvidia.com>
Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>
Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>