Handle uninitialized resources in structs
Fixes #8443

In order to support some legacy shaders, we will allow
zero-initialization of structs that contain resource types as fields.
The resource variables get initialized to a Poison value. We also
issue a diagnostic warning.

The expectation here is that (a) the correct value will be assigned in
SSA, or (b) if there's a branch scope where it remains initialized to
Poison, the value will subsequently not be used. (The latter case is
the one that we needed to implement this change for.)