fix bug in shared memory pointer (#9105)
Close #9110.

Fix the issue that we could generate invalid spirv code for shared
memory.
Slang supports using shared memory as a function parameter, however if
there is pointer offset operation inside the function, it will generate
invalid code.
e.g.
```
void foo(Ptr<....., AddressSpace::GroupShared> ptr)
{
    ptr[2] = xxx;
    or xxx = ptr[2];          // theses code will generate invalid code
}
```

The reason is that, our currently logic is to convert the pointer of
shared memory into a `Function` storage pointer, and pass it to a
function as the spirv restriction that the function call require the
pointer operand as a memory object.
This will bring a problem that for pointer offset operation like
OpPtrAccessChain, it will violate the rule because the storage class of
base pointer and result pointer type of this instruction is the
different (base is `Function` while result is `Workgroup`).

This conversion logic is added in the spirv legalization pass. However,
this restriction is removed when using Variable pointer capability and
when the pointer is `Workgroup` storage class. So we can skip this
conversion when the pointer has `GroupShared` address space.

---------

Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>