Tweak logic for SPIRV max ID bounds (#9103)
Fixes #9086

Increase the initial max ID to a much larger value. In optimization
modes where the CompactID pass isn't already run, manually apply that
pass if the ID bound of the final SPIRV is higher than
kDefaultMaxIdBound.

This gives the optimizer more room to maneuver,
and in low-optimization cases helps avoids invalid SPIRV if the
optimization result ended up with IDs beyond kDefaultMaxIdBound.

---------

Co-authored-by: slangbot <ellieh+slangbot@nvidia.com>
Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>
Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>