lldb python formatters: add object validation
When providing formatters for types in lldb, we need to iterate
through the types to find the appropriate data. If we attempt to
iterate into a leaf of an object that isn't valid (e.g., into a
nullptr) then python will generate a warning/error message on the
command line. If you're using lldb on the command line, and
particularly when using the curses-based gui, this results in visible
corruption.

This change validates each step along the way to avoid this issue.

(Coded by Claude)