Improve the IR uninitialized value diagnostic
During the IR uninitialized values pass, there is a diagnostic warning
that is generated for using an uninitialized value. If the source
value is a poison value, then the name is missing from the variable,
resulting in a confusing error message.

This change adds an alternate version of that warning message that
includes the type information whenever the name is unavailable.