Handle a file open attempt on a directory
This commit handles the case where file read attempt is on a directory
rather than a file. The issue descrbies that if a directory is read as a
file, the file size value goes wrong and the memory allocation will
fail.

The issue was reported on Linux and I was not reproducible on Windows as
described.