Fix default implement method in generic interface
Close #9202.

The root cause of this failure is that there is a specialize
inst, where we provide a generic type to a witness type.
generic %3 : %6
{
    witness_table_entry(%IWOTensorx5Fx5Fx5Fslangpyx5Fstore,
      specialize(specialize(%25, %T9, %D9), specialize(%4, %T9, %D9), %3))
}

%3 here in the specialize inst is to provide the subtypewitness, so it
should not be a generic, otherwise this specialize is just invalid.

The reason for this problem is that in frontend, when we check the
conformance of a struct, we will check all the inherianceDecl of a
struct, and for each inherianceDecl we will for a subtypewitness.
However, for the declRef field of the subtypewitness, we shouldn't
use directDeclRef, because there won't be any substitution happens
when it's a generic. So instead, we will form a GenericAppDeclRef
if the type of inherianceDecl is a generic.